requirements.txt
# Core dependencies
yara-python>=4.0.0
psutil>=5.8.0
capstone>=4.0.0
PyYAML>=6.0

# Cloud storage dependencies
boto3>=1.26.0
azure-storage-blob>=12.0.0
google-cloud-storage>=2.0.0

# Google Drive API dependencies
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Additional utilities
requests>=2.28.0

# SIEM/ELK integration
elasticsearch>=8.0.0

# PDF reporting
reportlab>=3.6.0

# Enhanced detection capabilities
pefile>=2023.2.7

# Timezone handling
pytz>=2023.3 